Crate sherwood

Structs

Config
FileHandle configuration

ConfigBuilder
Builds configuration for a file handle

Doc
Document which was fully retrieved from db

Error
ForestDB error

FileHandle
ForestDB file handle

FileHandleInfo
Provides stats for file handle

Iterator
Iterator over key-value store

IteratorOptions
Iterator options

KvHandle
ForestDB key value store

Meta
Meta information about a document key, seq_num, offset and meta itself

StoreConfig
Configuration of KvHandle

Enums

CommitOptions
Commit options

Durability
Durability options

IsolationLevel
Transaction isolation level

Location
Specifies how to search for document

SeekOptions
Specifies behavior of iterator if requested key is not found

Constants

NONE
Empty flags for iterator

NO_DELETES
Skip deleted documents

SKIP_MAX_KEY
Exclude range maximum value

SKIP_MIN_KEY
Exclude range minimum value

Traits

DocBody
Trait for accessing document body

DocKey
Trait for accessing key data from a document representation

DocMeta
Trait for accessing document metadata

FromBytes
Trait to convert value from a byte slice to a corresponding type

KeyRange
Trait which allows to create custom key range iterators

SeqRange
Trait which allows to create custom seq num iterators

Functions

init
Initializes ForestDB. Usually is called automatically

shutdown
Forces ForestDB shutdown: closing everything and terminating compactor thread

Type Definitions

FdbResult
ForestDB result

Snapshot
Read-only key-value handle

Store
Read-write key-value handle