Enum sherwood::SeekOptions

pub enum SeekOptions {
    Higher,
    Lower,
}

Specifies behavior of iterator if requested key is not found

Variants

Higher

If seek key does not exist return the next sorted key higher than it

Lower

If seek key does not exist return the previous sorted key lower than it