Struct sherwood::ConfigBuilder

pub struct ConfigBuilder {
    // some fields omitted
}

Builds configuration for a file handle

Methods

impl ConfigBuilder

fn new() -> ConfigBuilder

fn cache_size(self, size: u64) -> ConfigBuilder

Sets cache size 0 to disable cache at all Global for all file handles

fn wal_threshold(self, size: u64) -> ConfigBuilder

Sets WAL threshold Local to file handle

fn read_only(self) -> ConfigBuilder

Opens file handle in read-only mode

fn durability(self, durability: Durability) -> ConfigBuilder

Change durability options Local to file handle

fn chunk_size(self, size: u16) -> ConfigBuilder

Chunk size that is used to build B+-tree at each level Default 8 bytes, min 4 bytes, max 64 bytes Local to file handle

fn block_size(self, size: u32) -> ConfigBuilder

Size of block that is a unit of IO operations. Default 4Kb, min 1Kb max 128Kb. Global across all files

fn build(self) -> Config